moved function to utils
